edesia_main/requirements.txt
appdirs==1.4.4
argon2-cffi==20.1.0
asgiref==3.2.10
astroid==2.4.2
atomicwrites==1.4.0
attrs==19.3.0
autopep8==1.5.4
backcall==0.2.0
black==19.10b0
certifi==2020.6.20
cffi==1.14.1
cfgv==3.2.0
chardet==3.0.4
click==7.1.2
colorama==0.4.3
coverage==5.1
cryptography==3.1
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.1
Django==3.0.8
django-allauth==0.42.0
django-autoslug==1.9.8
django-cors-headers==3.5.0
django-countries==6.1.3
django-coverage-plugin==1.8.0
django-crispy-forms==1.9.2
django-debug-toolbar==2.2
django-environ==0.4.5
django-extensions==2.2.9
django-filter==2.3.0
django-model-utils==4.0.0
django-rest-knox==4.1.0
django-stubs==1.5.0
django-test-plus==1.4.0
djangorestframework==3.11.1
factory-boy==2.12.0
Faker==4.1.1
filelock==3.0.12
flake8==3.8.2
identify==1.4.25
idna==2.10
ipdb==0.13.2
ipython==7.17.0
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.17.2
lazy-object-proxy==1.4.3
Markdown==3.2.2
mccabe==0.6.1
more-itertools==8.4.0
mypy==0.770
mypy-extensions==0.4.3
nodeenv==1.4.0
oauthlib==3.1.0
packaging==20.4
parso==0.7.1
pathspec==0.8.0
pickleshare==0.7.5
Pillow==7.2.0
pluggy==0.13.1
pre-commit==2.4.0
prompt-toolkit==3.0.5
psycopg2-binary==2.8.5
py==1.9.0
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
Pygments==2.6.1
pylint==2.5.3
pylint-django==2.0.15
pylint-plugin-utils==0.6
pyparsing==2.4.7
pytest==5.4.2
pytest-django==3.9.0
pytest-sugar==0.9.3
python-dateutil==2.8.1
python-slugify==4.0.1
python3-openid==3.2.0
pytz==2020.1
PyYAML==5.3.1
regex==2020.7.14
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.3.1
termcolor==1.1.0
text-unidecode==1.3
toml==0.10.1
traitlets==4.3.3
typed-ast==1.4.1
typing-extensions==3.7.4.2
Unidecode==1.1.1
urllib3==1.25.10
virtualenv==20.0.30
wcwidth==0.2.5
Werkzeug==1.0.1
whitenoise==5.1.0
wrapt==1.12.1
